requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
crispy-bootstrap5==0.6
Django==4.0.3
django-ckeditor==6.2.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-js-asset==2.0.0
django-mptt==0.13.4
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
Pillow==9.0.1
psycopg2-binary==2.9.3
sqlparse==0.4.2
